Restructure publish
