+++ setup.py
@@ -56,6 +56,7 @@ setuptools.setup(
         "console_scripts": [
             "saib-tf = simple_ai_benchmarking.entrypoints:run_tf_benchmarks",
             "saib-pt = simple_ai_benchmarking.entrypoints:run_pt_benchmarks",
+            "saib-pub = simple_ai_benchmarking.entrypoints:publish",
         ]
     },
 )
