+++ simple_ai_benchmarking/workloads/tensorflow_workload.py
@@ -91,7 +91,7 @@ class TensorFlowKerasWorkload(AIWorkload):
         return tf.__version__
 
     def _get_ai_framework_extra_info(self) -> str:
-        extra_info = ""
+        extra_info = "N/A"
         if "AI_FRAMEWORK_EXTRA_INFO_TF" in os.environ:
             extra_info = os.environ["AI_FRAMEWORK_EXTRA_INFO_TF"]
         return extra_info
